Fix keyboard bindings for chain list
