+++ ropa/gui/controller/search_list_controller.py
@@ -13,5 +13,3 @@ class SearchListController(ListWidgetController):
             qg.QAbstractItemView.ScrollPerPixel)
         self.widget.setItemDelegate(
             HTMLDelegate(self.widget))
-        self.widget.keyPressEvent = self.key_press_event
-        self.widget.keyReleaseEvent = self.key_release_event
